+++ examples/hello_hardpy/test_simple.py
@@ -1,2 +1,4 @@
-def test_one():
+import hardpy
+
+def test_one(hardpy_start_args: dict):
     assert True